+++ core-rollback.php
@@ -12,7 +12,7 @@
  * Plugin Name:       Core Rollback
  * Plugin URI:        https://github.com/afragen/core-rollback
  * Description:       Seamless rollback of WordPress Core to latest release or any outdated, secure release using the Core Update API and core update methods.
- * Version:           0.5.0
+ * Version:           0.5.0.1
  * Author:            Andy Fragen
  * License:           MIT
  * Domain Path:       /languages
@@ -26,14 +26,13 @@
 
 namespace Fragen\Rollback;
 
+/*
+ * Exit if called directly.
+ */
+if ( ! defined( 'WPINC' ) ) {
+	die;
+}
+
 require_once __DIR__ . '/vendor/autoload.php';
 
-add_action(
-	'init',
-	function() {
-		load_plugin_textdomain( 'core-rollback' );
-	}
-);
-
-( new Core() )->load_hooks();
-( new Settings() )->load_hooks();
+( new Bootstrap() )->run();