+++ CHANGES.md
@@ -1,4 +1,5 @@
 [ unreleased ]
+* add `Class Bootstrap` to intiate process
 * updated instructions, etc
 * add assets, screenshots, banners, icon
 * add `readme.txt` and clean up `create_admin_page()`